Adding the website logo to the sidebar, improving the style of the sidebar options (home page, syllabus), renaming components
