+++ src/index.css
@@ -84,10 +84,3 @@ html {
 body {
   font-family: "Albert Sans";
 }
-
-main {
-  max-width: 90rem;
-  padding: 1rem 2rem;
-  margin: auto;
-  min-height: 100vh;
-}
