+++ src/layouts/StudentLayout.jsx
@@ -1,13 +1,13 @@
 import { Header } from "../domains/core/componentes/organism/Header";
 import { Outlet } from "react-router-dom";
-import { SidebarStudent } from "../shared/components/organisms/SidebarStudent";
+import { StudentSidebar } from "../shared/components/organisms/StudentSidebar";
 
 export const StudentLayOut = () => {
   return (
     <>
-      <div id="teahcerWorkspaceLayout" className="min-h-screen flex flex-row">
-        <SidebarStudent />
-        <div className="mainContent  w-full">
+      <div className="min-h-screen flex flex-row">
+        <StudentSidebar />
+        <div className=" w-full flex flex-col">
           <Header complete={false}></Header>
           <Outlet />
         </div>
