+++ src/shared/components/organisms/StudentSidebar.jsx
@@ -2,8 +2,8 @@ import { useState } from "react";
 import { Icon } from "../atoms/Icon";
 import { SidebarRow } from "../molecules/SidebarRow";
 import { Link } from "react-router-dom";
-
-export const SidebarStudent = () => {
+import { SidebarLogo } from "../molecules/SidebarLogo";
+export const StudentSidebar = () => {
   const [open, setOpen] = useState(false);
 
   return (
@@ -13,13 +13,14 @@ export const SidebarStudent = () => {
       className="w-[3.7rem] "
     >
       <div className="relative h-full">
-        <div className="absolute left-0 top-0 h-full z-[10] border-r border-gray-400">
+        <div className="absolute left-0 top-0 h-full z-[20] border-r border-gray-400">
           <div className="absolute top-0 left-0 h-full z-20 transition-all duration-[400ms] ease-[cubic-bezier(0.2,0,0.38,0.9)]">
             <nav
               className={`${
                 open ? "w-52" : "w-[3.7rem]"
               } h-full flex flex-col  bg-white border-r border-gray-300 shadow-md shadow-gray-300 sticky top-0 max-h-screen overflow-y-auto overflow-x-hidden transition-all duration-300`}
             >
+              <SidebarLogo open={open} route={"/student/dashboard"} />
               <SidebarRow
                 name={"Dashboard"}
                 route={"/student/dashboard"}
